attcpo

compute the Average Treatment Effect on the Treated Conditional on the previous outcome (ATT-CPO)

attcpo(formla, t, tmin1, tmin2, tname, data, idname, Y0tqteobj, h = NULL,
  yseq = NULL, yseqlen = 100, se = TRUE, iters = 100,
  method = "level")

Arguments

formla
e.g. y ~ treat

t
the last time period

tmin1
the middle time period

tmin2
the first time period

tname
the name of the column containing time periods in the data

data
a data.frame

idname
the name of the column containing an individual identifier over time

Y0tqteobj
a qte object (from the qte package) containing the the counterfactual distribution of untreated potential outcomes for the treated group

h
optional bandwidth

yseq
optional sequence of y values, default is to use all unique yvalues in the data, though this can increase computation time

yseqlen
optional length of y values to use, aids in automatically generating yseq if desired

se
whether or not to compute standard errors

iters
how many bootstrap iterations to use if computing standard errors; default is 100.

method
should be either "levels" or "rank"; whether to compute the ATT-CPO using based on the levels of Y0tmin1 or the ranks of Y0tmin1; "levels" is the default.

Value

att-cpo