cia.bounds

Bounds on the distribution of the treatment effect and on the quantile of the treatment effect under a conditional independence assumption (cia). It takes in a data.frame, that should indicate whether or not an individual is treated; separates individuals into a treated and untreated group; runs distribution or quantile regression to estimate the conditional distributions; then computes bounds on the DoTT or QoTT.

cia.bounds(
  formla,
  xformla = ~1,
  data,
  delt.seq,
  y.seq = NULL,
  firststep = c("dr", "qr", "ll"),
  link = "logit",
  se = FALSE,
  bootiters = 100,
  cl = 1,
  alp = 0.05,
  ...
)

Arguments

formla

y ~ d

xformla

~ x1 + x2

data

the name of the data.frame. It should be in "long" format rather than "wide" format.

delt.seq

a vector of values to compute the distribution of the treatment effect for

y.seq

a vectof of values to compute first-step distributions over (this is currently not used as it is computed internally)

firststep

whether to use distribution regression ("dr"), quantile regression ("qr"), or local linear distribution regression ("ll") for the first step estimation of condtional distributions

link

optional argument to pass to distreg for which link function to use when running distribution regressions

se

whether or not to compute standard errors (if TRUE, they are computed using the bootstrap

bootiters

if computing standard errors using the bootstrap, how many bootstrap iterations to use

cl

if computing standard errors using the bootstrap, how many cores to use in parallel computation (default is 1)

alp

significance level for confidence intervals

...

whatever extra arguments need to be passed to Y0tmethod