compute.attcpo

does the heavy lifting for computing the attcpo

compute.attcpo(
  Y1t,
  Y0tmin1,
  Y0tmin2,
  Y0tqteobj,
  h = NULL,
  ytmin1seq = NULL,
  yseqlength = NULL,
  se = TRUE,
  iters = 100,
  method = "level"
)

Arguments

Y1t

treated group outcomes in period t

Y0tmin1

treated group outcomes in period t-1

Y0tmin2

treated group outcomes in period t-2

Y0tqteobj

qte object to get counterfactual quantiles from

h

optional bandwidth

ytmin1seq

optional sequence of values to compute attcpo(y') for

yseqlength

optional length of sequence of values to compute attcpo(y') for; default is 100

se

boolean for whether or not to compute standard errors

iters

number of bootstrap iterations to use to compute standard errors

method

can be "level" or "rank", whether the conditional expectation is based on the level of Y0tmin1 or its rank