compute.csa.bounds

Compute bounds on the distribution and quantile of the treatment effect as given in Callaway (2017) under the copula stability assumption and when a first step estimator of the counterfactual distribution of untreated potential outcomes for the treated group is available.

compute.csa.bounds(
  formla,
  t,
  tmin1,
  tmin2,
  tname,
  idname,
  data,
  delt.seq,
  y.seq,
  Y0tqteobj,
  F.y0 = NULL,
  F.y1 = NULL,
  h = NULL,
  xformla = ~1,
  firststep = c("dr", "qr", "ll"),
  cl = 1
)

Arguments

formla
outcomevar ~ treatmentvar

t
the 3rd period

tmin1
the 2nd period

tmin2
the 1st period

tname
the name of the column containing periods

idname
the name of the column containing ids

data
a panel data frame

delt.seq
the possible values to compute bounds on the distribution of the treatment effect for

y.seq
the possible values for y to take

Y0tqteobj
a previously computed first step estimator of the distribution of counterfactual outcomes for the treated group

F.y0
(optional) pre-computed distribution of counterfactual untreated outcomes for the treated group

F.y1
(optional) pre-computed distribution of treated outcomes for the treated group

h
optional bandwidth when using local linear regression

xformla
a formula for which covariates to use

firststep
whether to use distribution regression ("dr"), quantile regression ("qr"), or local linear distribution regression ("ll") for the first step estimation of condtional distributions

cl
(optional) number of multi-cores to use

Value

csaboundsobj