csa.bounds

Compute bounds on the distribution and quantile of the treatment effect as given in Callaway (2017) under the copula stability assumption and when a first step estimator of the counterfactual distribution of untreated potential outcomes for the treated group is available.

csa.bounds(delt.seq, y.seq, Y1t, Y0tmin1, Y0tmin2, Y0tqteobj, F.y0 = NULL,
  F.y1 = NULL, h = NULL, method = c("level", "rank"))

Arguments

delt.seq
the possible values to compute bounds on the distribution of the treatment effect for

y.seq
the possible values for y to take

Y1t
vector of outcomes for the treated group in period t

Y0tmin1
vector of outcomes for the treated group in period t-1

Y0tmin2
vector of outcomes for the treated group in period t-2

Y0tqteobj
a qte object for obtaining the counterfactual distribution of untreated potential outcomes for the treated group in period t

F.y0
(optional) pre-computed distribution of counterfactual untreated outcomes for the treated group

F.y1
(optional) pre-computed distribution of treated outcomes for the treated group

h
optional bandwidth

method
"level" or "rank" determining whether method should be used conditional on ytmin1 or the rank of ytmin1

Value

csaboundsobj

Examples

data(displacements)
head(displacements)

#>       id treat year     learn
#> 4437  28     1 2011 10.043249
#> 4493  84     1 2011  9.472705
#> 4519 110     1 2011 10.621327
#> 4653 244     1 2011 11.082143
#> 4654 245     1 2011 10.043249
#> 4703 294     1 2011  9.392662