ggCSABounds

plot bounds on the quantile of the treatment effect using ggplot2

ggCSABounds(
  csaboundsobj,
  tau = seq(0.05, 0.95, 0.05),
  csalabel = "CSA Bounds",
  wdbounds = FALSE,
  wdlabel = "WD Bounds",
  otherdist1 = NULL,
  otherdist1label = "CS RI",
  otherdist2 = NULL,
  otherdist2label = "Panel RI",
  plotSE = TRUE,
  colors = NULL,
  legend = TRUE
)

Arguments

csaboundsobj

an object returned from the csa.bounds method

tau

vector of values between 0 and 1 to plot quantiles for

csalabel

label for bounds under copula stability assumption

wdbounds

boolean whether or not to also plot Williamson-Downs bounds

wdlabel

label for worst case bounds

otherdist1

optional ecdf of the distribution of the treatment effect under cross sectional rank invariance

otherdist1label

label for other distribution 1

otherdist2

optional ecdf of the distribution of the treatment effect under panel rank invariance

otherdist2label

label for other distribution 2

plotSE

whether or not to plot standard errors (default is TRUE)

colors

optional vector of colors to use in place of ggplot's default scheme

legend

whether or not to include a legend (default is TRUE)